fix: correct type assertion from Float32Array to Float64Array
HEAPF64 indicates 64-bit floating point values, so the type assertion
should be Float64Array, not Float32Array. This improves type safety
as suggested in the code review.

+++ src/modifiers/colormodifier.ts
@@ -157,7 +157,7 @@ class ColorModifier extends Modifier {
     const perAtomArray = input.wasm.HEAPF64.subarray(
       perAtomDataPtr,
       perAtomDataPtr + output.particles.count,
-    ) as Float32Array;
+    ) as Float64Array;
     // Calculate min/max using a loop to avoid stack overflow for large arrays
     let minValue = Infinity;
     let maxValue = -Infinity;
